Prevented NaN propagation in ZMatrix operations
Added debugging features using pseudo static toggles

+++ V4.0/ZDK/src/nl/zeesoft/zdk/neural/TestSet.java
@@ -73,7 +73,7 @@ public class TestSet implements JsAble {
 			averageError = averageError / total;
 		}
 		if (tests.size()>0 && averageLoss>0) {
-			averageLoss = averageLoss / tests.size();
+			averageLoss = averageLoss / (float) tests.size();
 		}
 	}
